requirements.txt
asgiref==3.8.1
dj-database-url==2.3.0
Django==5.1.7
django-cors-headers==4.7.0
djangorestframework==3.15.2
gunicorn==23.0.0
packaging==24.2
psycopg2-binary==2.9.10
python-decouple==3.8
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
whitenoise==6.9.0
